+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/FVPVideoPlayerPlugin.m
@@ -90,10 +90,46 @@
   return textureId;
 }
 
+// This function, although slightly modified, is also in camera_avfoundation.
+// Both need to do the same thing and run on the same thread (for example main thread).
+// Do not overwrite PlayAndRecord with Playback which causes inability to record
+// audio, do not overwrite all options.
+// Only change category if it is considered an upgrade which means it can only enable
+// ability to play in silent mode or ability to record audio but never disables it,
+// that could affect other plugins which depend on this global state. Only change
+// category or options if there is change to prevent unnecessary lags and silence.
+#if TARGET_OS_IOS
+static void upgradeAudioSessionCategory(AVAudioSessionCategory requestedCategory,
+                                        AVAudioSessionCategoryOptions options,
+                                        AVAudioSessionCategoryOptions clearOptions) {
+  NSSet *playCategories = [NSSet
+      setWithObjects:AVAudioSessionCategoryPlayback, AVAudioSessionCategoryPlayAndRecord, nil];
+  NSSet *recordCategories =
+      [NSSet setWithObjects:AVAudioSessionCategoryRecord, AVAudioSessionCategoryPlayAndRecord, nil];
+  NSSet *requiredCategories =
+      [NSSet setWithObjects:requestedCategory, AVAudioSession.sharedInstance.category, nil];
+  BOOL requiresPlay = [requiredCategories intersectsSet:playCategories];
+  BOOL requiresRecord = [requiredCategories intersectsSet:recordCategories];
+  if (requiresPlay && requiresRecord) {
+    requestedCategory = AVAudioSessionCategoryPlayAndRecord;
+  } else if (requiresPlay) {
+    requestedCategory = AVAudioSessionCategoryPlayback;
+  } else if (requiresRecord) {
+    requestedCategory = AVAudioSessionCategoryRecord;
+  }
+  options = (AVAudioSession.sharedInstance.categoryOptions & ~clearOptions) | options;
+  if ([requestedCategory isEqualToString:AVAudioSession.sharedInstance.category] &&
+      options == AVAudioSession.sharedInstance.categoryOptions) {
+    return;
+  }
+  [AVAudioSession.sharedInstance setCategory:requestedCategory withOptions:options error:nil];
+}
+#endif
+
 - (void)initialize:(FlutterError *__autoreleasing *)error {
 #if TARGET_OS_IOS
   // Allow audio playback when the Ring/Silent switch is set to silent
-  [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback error:nil];
+  upgradeAudioSessionCategory(AVAudioSessionCategoryPlayback, 0, 0);
 #endif
 
   [self.playersByTextureId
@@ -204,11 +240,11 @@
   // AVAudioSession doesn't exist on macOS, and audio always mixes, so just no-op.
 #else
   if (mixWithOthers) {
-    [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback
-                                     withOptions:AVAudioSessionCategoryOptionMixWithOthers
-                                           error:nil];
+    upgradeAudioSessionCategory(AVAudioSession.sharedInstance.category,
+                                AVAudioSessionCategoryOptionMixWithOthers, 0);
   } else {
-    [[AVAudioSession sharedInstance] setCategory:AVAudioSessionCategoryPlayback error:nil];
+    upgradeAudioSessionCategory(AVAudioSession.sharedInstance.category, 0,
+                                AVAudioSessionCategoryOptionMixWithOthers);
   }
 #endif
 }