+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.9.17+6
+
+* Fixes changing global audio session category to be collision free across plugins.
+
 ## 0.9.17+5
 
 * Adds ability to use any supported FPS and fixes crash when using unsupported FPS.