+++ packages/video_player/video_player_avfoundation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.6.6
+
+* Fixes changing global audio session category to be collision free across plugins.
+
 ## 2.6.5
 
 * Bugfix to allow the audio-only HLS (.m3u8) on iOS.